+++ arkworks_adapter/src/integration_tests.rs
@@ -1,7 +1,4 @@
-use ark_relations::{
-    lc,
-    r1cs::{ConstraintSynthesizer, ConstraintSystem, ConstraintSystemRef, SynthesisError},
-};
+use ark_relations::r1cs::{ConstraintSynthesizer, ConstraintSystem};
 
 use pinocchio_vm::{
     circuits::qap::QuadraticArithmeticProgram as Qap,
@@ -15,6 +12,7 @@ use pinocchio_vm::{
 
 use crate::{
     arkworks_cs_to_pinocchio_r1cs, arkworks_io_and_witness_to_pinocchio_io_and_witness, fq5::Fq,
+    test_utils::PinocchioPaperExampleCircuit,
 };
 
 #[test]
@@ -52,34 +50,4 @@ fn create_proof_from_arkworks_and_verify_it() {
     let accepted = verifier::verify(&vk, &proof, &io);
 
     assert!(accepted);
-}
-
-//TO DO: Move this module to a common library
-pub struct PinocchioPaperExampleCircuit {
-    /// Public input
-    pub a: Fq,
-    /// Private input
-    pub b: Fq,
-    pub c: Fq,
-    pub d: Fq,
-}
-
-impl ConstraintSynthesizer<Fq> for PinocchioPaperExampleCircuit {
-    fn generate_constraints(self, cs: ConstraintSystemRef<Fq>) -> Result<(), SynthesisError> {
-        let a = cs.new_input_variable(|| Ok(self.a))?;
-        let b = cs.new_input_variable(|| Ok(self.b))?;
-        let c = cs.new_input_variable(|| Ok(self.c))?;
-        let d = cs.new_input_variable(|| Ok(self.d))?;
-
-        let e = cs.new_witness_variable(|| Ok(self.c * self.d))?;
-        cs.enforce_constraint(lc!() + c, lc!() + d, lc!() + e)?;
-
-        let calculated_result = self.c * self.d * (self.a + self.b);
-
-        let result = cs.new_input_variable(|| Ok(calculated_result))?;
-
-        cs.enforce_constraint(lc!() + a + b, lc!() + e, lc!() + result)?;
-
-        Ok(())
-    }
 }
